add href on detail tab to match
